Add ErrorHandler with supporting classes and update settings
The commit introduces an ErrorHandler class that catches and logs exceptions and PHP errors. It also includes a variety of new exception types, a settings class and related factories. A new service, PilotClientService, has been created to send exceptions to a server. The associated composer and ignore files have also been updated.